backend/requirements.txt
Flask>=3.1.2
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
openai>=1.50.0
PyJWT>=2.8.0
pandas
python-docx
PyPDF2
azure-cosmos
